+++ Sources/SafariWalletCore/Client/Alchemy/AlchemyClient.swift
@@ -131,4 +131,5 @@ extension AlchemyClient {
         }
         return allowance
     }
+   
 }
