+++ Sources/SafariWalletCore/Client/JsonRpc/JsonRpcClient.swift
