+++ Sources/SafariWalletCore/Client/JsonRpc/JsonRpcModels/JsonRpcError.swift
