+++ Sources/SafariWalletCore/Client/JsonRpc/JsonRpcModels/JsonRpcRequest.swift
