+++ Sources/SafariWalletCore/Client/JsonRpc/JsonRpcModels/JsonRpcResponse.swift
